Fixes #212, rename test
