+++ sqlg-test/src/main/java/org/umlg/sqlg/AllTest.java
@@ -27,7 +27,7 @@ import org.umlg.sqlg.test.rollback.TestRollback;
 import org.umlg.sqlg.test.schema.*;
 import org.umlg.sqlg.test.topology.TestTopologyChangeListener;
 import org.umlg.sqlg.test.topology.TestTopologyDelete;
-import org.umlg.sqlg.test.topology.TestTopologyDeleteBatch;
+import org.umlg.sqlg.test.topology.TestTopologyDeleteSpecific;
 import org.umlg.sqlg.test.topology.TestTopologyMultipleGraphs;
 import org.umlg.sqlg.test.topology.TestTopologyUpgrade;
 import org.umlg.sqlg.test.topology.TestValidateTopology;
@@ -171,7 +171,7 @@ import org.umlg.sqlg.test.vertexstep.localvertexstep.*;
         TestBatchNormalUpdateDateTimeArrays.class,
         TestTopologyChangeListener.class,
         TestTopologyDelete.class,
-        TestTopologyDeleteBatch.class
+        TestTopologyDeleteSpecific.class
 })
 public class AllTest {
 }
